requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.62.1
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-crispy-forms==2.1
django-mptt==0.16.0
django-recaptcha==4.0.0
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
django-summernote==0.8.20.0
idna==3.7
isodate==0.6.1
kombu==5.3.7
lxml==5.2.1
oauthlib==3.2.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
python3-saml==1.16.0
redis==5.0.4
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
xmlsec==1.3.14
